feat: add activity session IDs to mission summary
