+++ web/app/routes/_layout.noctis-team/components/noctis-team-screen.tsx
@@ -1,6 +1,6 @@
-import { Archive, Check, Ellipsis, History, Pencil, Plus, RotateCcw, X } from "lucide-react";
-import { memo, useCallback, useEffect, useMemo, useState } from "react";
-import { NavLink, useMatch, useNavigate, useParams } from "react-router";
+import { Ellipsis, History, Plus } from "lucide-react";
+import { useCallback, useEffect, useMemo, useState } from "react";
+import { useMatch, useNavigate, useParams } from "react-router";
 import { toast } from "sonner";
 import { WorkspaceLaunchActions } from "@/components/workspace-launch-actions";
 import { Button } from "@/components/ui/button";
@@ -27,7 +27,6 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
-import { Textarea } from "@/components/ui/textarea";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import {
   type MissionResumePayload,
@@ -35,6 +34,7 @@ import {
   useAgentSession,
 } from "@/hooks/use-agent-session";
 import { useProjectRegistry } from "@/hooks/use-project-registry";
+import { useSessionStatusFeed } from "@/hooks/use-session-status-feed";
 import { useVSCodePreferences } from "@/hooks/use-vscode-preferences";
 import type { AppLanguage } from "@/lib/app-language.server";
 import { normalizeContextProjectIds } from "@/lib/execution-context";
@@ -54,12 +54,14 @@ import type {
   MissionOutputSummary,
   MissionSurfaceId,
 } from "@/lib/types/mission";
+import { isMissionSummaryRunning } from "@/lib/mission-list-running-state";
 import { cn } from "@/lib/utils";
 import type { PromptPart } from "@/lib/prompt-parts";
 import type { MessageInfo } from "@/routes/_layout.opencode.session.$id/types";
 import { BanterLog } from "./banter-log";
 import { ChatArea } from "./chat-area";
 import { LunafreyaStatusPanel, type LunafreyaFacetOption } from "./lunafreya-status-panel";
+import { MissionHistoryItem } from "./mission-history-item";
 import { MissionActivityLog } from "./mission-activity-log";
 import {
   getMissionOutputKey,
@@ -82,162 +84,6 @@ type BulkMissionDialogState = {
 } | null;
 
 type InspectorTab = "banter" | "activity" | "outputs";
-
-type MissionHistoryItemProps = {
-  mission: MissionSummary;
-  routeBase: string;
-  isActive: boolean;
-  isArchivedView: boolean;
-  isEditing: boolean;
-  isArchivePending: boolean;
-  isArchiveDisabled: boolean;
-  isRenaming: boolean;
-  onBeginRename: (mission: MissionSummary) => void;
-  onArchiveAction: (mission: MissionSummary, action: "archive" | "restore") => void;
-  onCancelRename: () => void;
-  onSubmitRename: (missionId: string, title: string) => void;
-};
-
-const MissionHistoryItem = memo(
-  ({
-    mission,
-    routeBase,
-    isActive,
-    isArchivedView,
-    isEditing,
-    isArchivePending,
-    isArchiveDisabled,
-    isRenaming,
-    onBeginRename,
-    onArchiveAction,
-    onCancelRename,
-    onSubmitRename,
-  }: MissionHistoryItemProps) => {
-    const [draftTitle, setDraftTitle] = useState(mission.title || "Untitled mission");
-
-    useEffect(() => {
-      if (isEditing) {
-        setDraftTitle(mission.title || "Untitled mission");
-      }
-    }, [isEditing, mission.title]);
-
-    if (isEditing) {
-      return (
-        <div className="rounded-xl border border-primary/30 bg-primary/8 p-3">
-          <Textarea
-            value={draftTitle}
-            onChange={(event) => setDraftTitle(event.target.value)}
-            rows={2}
-            disabled={isRenaming}
-            className="min-h-14 resize-none bg-transparent text-xs"
-          />
-          <div className="mt-2 flex items-center justify-end gap-1">
-            <Button
-              type="button"
-              variant="ghost"
-              size="icon"
-              className="h-7 w-7"
-              onClick={onCancelRename}
-              disabled={isRenaming}
-              title="Cancel rename"
-            >
-              <X className="h-3.5 w-3.5" />
-            </Button>
-            <Button
-              type="button"
-              variant="ghost"
-              size="icon"
-              className="h-7 w-7"
-              onClick={() => onSubmitRename(mission.missionId, draftTitle)}
-              disabled={isRenaming || !draftTitle.trim()}
-              title="Save title"
-            >
-              <Check className="h-3.5 w-3.5" />
-            </Button>
-          </div>
-        </div>
-      );
-    }
-
-    return (
-      <div
-        className={cn(
-          "group relative w-full min-w-0 max-w-full overflow-hidden rounded-xl border p-3 transition-colors",
-          isActive ? "border-primary/40 bg-primary/10" : "border-border/50 bg-card/40 hover:bg-card/70"
-        )}
-      >
-        <NavLink
-          aria-label={`Open mission ${mission.title}`}
-          className="absolute inset-0 rounded-xl focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-inset focus-visible:ring-primary/40"
-          to={buildMissionPath(mission.missionId, routeBase)}
-        />
-        <div className="grid w-full min-w-0 max-w-full grid-cols-[minmax(0,1fr)_auto] items-start gap-2 overflow-hidden">
-          <div className="pointer-events-none min-w-0 overflow-hidden">
-            <span className="block min-w-0 font-semibold text-sm leading-5 line-clamp-2 wrap-break-word">
-              {mission.title}
-            </span>
-            <div className="mt-2 flex min-w-0 items-center gap-2">
-              <span className="max-w-24 shrink-0 truncate rounded-full border border-border/50 px-2 py-0.5 font-mono text-[9px] uppercase tracking-widest text-muted-foreground/70">
-                {mission.status}
-              </span>
-              <p className="min-w-0 flex-1 truncate text-right font-mono text-[9px] uppercase tracking-widest text-muted-foreground/40">
-                {new Date(mission.updatedAt).toLocaleString("en-US", {
-                  year: "numeric",
-                  month: "2-digit",
-                  day: "2-digit",
-                  hour: "2-digit",
-                  minute: "2-digit",
-                  hour12: false,
-                })}
-              </p>
-            </div>
-          </div>
-          <div className="relative z-10 flex items-center gap-1">
-            {!isArchivedView ? (
-              <Button
-                type="button"
-                variant="ghost"
-                size="icon"
-                className={cn(
-                  "h-6 w-6 shrink-0 transition-[opacity,color,background-color]",
-                  "bg-background/30 text-foreground/70 opacity-0",
-                  "group-hover:opacity-100 hover:bg-accent hover:text-foreground",
-                  "focus-visible:opacity-100"
-                )}
-                onClick={() => onBeginRename(mission)}
-                title="Rename mission"
-              >
-                <Pencil className="h-3 w-3" />
-              </Button>
-            ) : null}
-            <Button
-              type="button"
-              variant="ghost"
-              size="icon"
-              className={cn(
-                "h-6 w-6 shrink-0 transition-[opacity,color,background-color]",
-                "bg-background/30 text-foreground/70 opacity-0",
-                "group-hover:opacity-100 hover:bg-accent hover:text-foreground",
-                "focus-visible:opacity-100"
-              )}
-              onClick={() => onArchiveAction(mission, isArchivedView ? "restore" : "archive")}
-              disabled={isArchivePending || isArchiveDisabled}
-              title={
-                isArchivedView
-                  ? "Restore mission"
-                  : isArchiveDisabled
-                    ? "Active mission cannot be archived while running"
-                    : "Archive mission"
-              }
-            >
-              {isArchivedView ? <RotateCcw className="h-3 w-3" /> : <Archive className="h-3 w-3" />}
-            </Button>
-          </div>
-        </div>
-      </div>
-    );
-  }
-);
 
 export interface NoctisTeamScreenProps {
   activeMissionId: string | null;
@@ -653,6 +499,8 @@ export function NoctisTeamScreen({
       setIsLoadingMissions(false);
     }
   }, [missionApiBase]);
+
+  const sessionStates = useSessionStatusFeed({ onSessionIdle: () => void loadMissions() });
 
   const loadMissionOutputs = useCallback(async () => {
     if (!effectiveMissionId) {
@@ -1228,6 +1076,7 @@ export function NoctisTeamScreen({
                       key={mission.missionId}
                       mission={mission}
                       routeBase={missionRouteBase}
+                      isRunning={isMissionSummaryRunning(mission, sessionStates)}
                       isActive={effectiveMissionId === mission.missionId}
                       isArchivedView={missionView === "archived"}
                       isEditing={editingMissionId === mission.missionId}
